+++ scripts/install_rvc_venv.py
@@ -10,13 +10,15 @@ if sys.platform == "win32": # Если операционная система -
     activate_script = "venv\\rvc_venv\\Scripts\\activate"
     commands_after_activation = [
         f"{activate_script} && pip install rvc-python",
-        f"{activate_script} && pip install torch==2.1.1+cu118 torchaudio==2.1.1+cu118 --index-url https://download.pytorch.org/whl/cu118"
+        f"{activate_script} && pip install openvoice-cli",
+        f"{activate_script} && pip install torch==1.13.1+cu117 torchaudio==0.13.1+cu117 --index-url https://download.pytorch.org/whl/cu117"
     ]
 else:  # For Unix-like systems
     activate_script = "venv\\rvc_venv\\bin\\activate"
     commands_after_activation = [
         f"source {activate_script} && pip install rvc-python",
-        f"source {activate_script} && pip install torch==2.1.1+cu118 torchaudio==2.1.1+cu118 --index-url https://download.pytorch.org/whl/cu118"
+        f"source {activate_script} && pip install openvoice-cli",
+        f"source {activate_script} && pip install torch==1.13.1+cu117 torchaudio==0.13.1+cu117 --index-url https://download.pytorch.org/whl/cu117"
     ]
 
 for command in commands_after_activation:
