+++ start_xtts_webui.bat
@@ -4,6 +4,6 @@ SET TEMP=temp
 
 call venv/scripts/activate
 
-python app.py --deepspeed
+python app.py --deepspeed --rvc
 
 pause
